test(common): move the interpolate_to_cell_center_vp test next to its stencil
The stencil moved to `common` but its test stayed in dycore, which is what
OngChia flagged on the port. It now sits beside the stencil, next to the
`_wp` sibling.

Its numpy reference goes to `model.testing.reference_funcs`: two dycore
advection tests import it, and test directories are not importable across
packages. Same route `compute_tangential_wind_numpy` took earlier on this
branch.
